Added the macro PHYSICAL_DEVICE_SCORE_SELECTION to CMakeLists.txt . Moved tinyobj and opencv headers and functions from vulkan_engine to main.cpp file. Adjusted the createSampledImage function to receive image data and dimensions instead of filename.
